requirements.txt
aiohttp==3.8.5
async_timeout==4.0.3
cogs==0.4.4
config==0.5.1
countdown==0.0.3
discord.py==2.3.2
python-dotenv==1.0.1
